Agregando funcionalidad de hacer solicitudes / eliminar solicitudes / consultar solicitudes / consultar el estado actual del culturest

+++ src/main/java/com/occe/controller/AlumnoController.java
@@ -35,7 +35,7 @@ public class AlumnoController {
     }
     
     @GetMapping("/plan-programa/{expediente}")
-    private PlanProgramaAlumno getPlanProgramaAlumno(@PathVariable("expediente") Long expediente){
+    private PlanProgramaAlumno getPlanProgramaAlumno(@PathVariable("expediente") Integer expediente){
         return alumnoService.getPlanPrograma(expediente);
     }
